+++ src/commands/release/index.ts
@@ -213,6 +213,6 @@ export default async function (config: OmniConfig, iterTactic?: {
     logTime('项目发布', true);
     handleReleaseSuc()();
   } catch (err) {
-    handleReleaseErr(`糟糕！发布过程发生了一点意外 (Oops! release process occured some accidents) \n👉  ${err}`)();
+    handleReleaseErr(`糟糕！发布过程发生了一点意外 (Oops! release process occured some accidents) \n👉  ${JSON.stringify(err)}`)();
   }
 }