+++ src/commands/dev/server.ts
@@ -85,10 +85,6 @@ function server ({
       logInfo('> Ready on local: ' + url_local);
       logInfo('> Ready on ip: ' + url_ip);
     });
-
-    (['SIGINT', 'SIGQUIT', 'SIGTERM'] as NodeJS.Signals[]).forEach((sig) => {
-      process.on(sig, () => process.exit(0));
-    });
   } catch (err) {
     logErr(err);
     process.exit(1);