+++ src/utils/index.ts
@@ -1,4 +1,3 @@
 export { default as logo } from './logo';
-export { default as module_path } from './module_path';
 export { default as signal } from './signal';
 export { getHandlers, handlerFactory } from './tackle_plugins';