+++ docs/CHANGELOG.md
@@ -4,7 +4,7 @@
 ### v1.2.16
 1. upgrade @omni-door/utils
 
-2. remove all `JSON.stringify` in the error-log output
+2. remove redundant `JSON.stringify` in the error-log output
 
 ### v1.2.15
 1. optimization exception