+++ docs/CHANGELOG.md
@@ -1,6 +1,9 @@
 # CHANGE LOG
 
 ## v1.2.x
+### v1.2.17
+1. [omni new] remove the logic which automatic transform to uppercase for the component-name's fist letter
+
 ### v1.2.16
 1. upgrade @omni-door/utils